Move fmt_baseline helper to utils
